chore: bundle small-mechanical-wins (CLAUDE.md paths, diagram move, etc.)
Eight low-risk cleanups grouped to amortise PR overhead:

- CLAUDE.md citations of src/utils/{tryCatch,responses,time,
  computeMapState}.mjs corrected to src/shared/utils/... locations.
- Moved src/shared/utils/diagram.mjs to src/app/docs/_diagram.mjs;
  10 importers + the unit test updated.
- umami.mjs: tryCatch + clearer log; inlined production hostname,
  dropped the unreachable getHostname() dev/staging branches.
- reloadGuard.mjs: flattened the two stacked conditionals into a
  single prevAttempts decision.
- responses.mjs: numeric range check instead of String.startsWith.
- Dropped numbered scaffolding comments from the campaign and
  rebroadcast routes.
- seed.mjs: collapsed `let files; files = data;` into a destructure.
- initializeEnvironmentVariables kept async deliberately with an
  in-file comment explaining the error semantics (caller relies on
  it for tryCatch wrapping; banned try/catch otherwise).

All 1299 unit tests pass. lint, typecheck, and build clean.
