+++ CLAUDE.md
@@ -73,7 +73,7 @@ Prefer these commands over manual git operations.
 
 ### Error Handling
 
-Use the `tryCatch` wrapper (`src/utils/tryCatch.mjs`). Do NOT use try/catch blocks.
+Use the `tryCatch` wrapper (`src/shared/utils/tryCatch.mjs`). Do NOT use try/catch blocks.
 
 ```js
 const { data, error } = await tryCatch(someAsyncOperation());
@@ -84,8 +84,8 @@ if (error) {
 
 ### API Routes
 
-- Use `errorResponse(code, start, error)` and `successResponse(code, start, data)` from `src/utils/responses.mjs`
-- Measure execution time with `roundedPerformanceTime(start)` from `src/utils/time.mjs`
+- Use `errorResponse(code, start, error)` and `successResponse(code, start, data)` from `src/shared/utils/api/responses.mjs`
+- Measure execution time with `roundedPerformanceTime(start)` from `src/shared/utils/time.mjs`
 
 ### Analytics Tracking
 
@@ -143,7 +143,7 @@ All visual properties use CSS custom properties defined in the Tailwind v4 `@the
 - **Node version:** mise pins node@24 (ships with npm 11 natively).
 - **Server actions:** Most utilities use `'use server'` directive.
 - **Shared utilities:** `formatNumber` (`src/shared/utils/format/formatNumber.mjs`) for compact numbers (25.0M, 1.2K — M suffix at 10M+, locale grouping below). `formatTimeAgo` (`src/shared/utils/format/formatTimeAgo.mjs`) for relative timestamps.
-- **Map state:** `computeMapState` (`src/utils/computeMapState.mjs`) computes galaxy map sector ownership. Sectors 1-10 from campaign `points`/`points_max`; region 11 (homeworld) from attack events only. **Critical:** live views must only pass active events — completed events are already in the score.
+- **Map state:** `computeMapState` (`src/shared/utils/game/computeMapState.mjs`) computes galaxy map sector ownership. Sectors 1-10 from campaign `points`/`points_max`; region 11 (homeworld) from attack events only. **Critical:** live views must only pass active events — use the `computeLiveMapState(data)` helper from the same module to keep the filter and the call together.
 - **On-demand season fetching:** `/archives` page derives SeasonSelector from current season number (not DB query). Missing seasons are backfilled from the official HD1 API on first request via `updateSeason()` (`src/update/season.mjs`) -- the same shared pipeline the worker runs every poll for the active season and the admin "Refresh" button triggers via `reseedSeason`. `updateSeason` writes `h1_season` (with inlined arrays) + `h1_status` + `h1_statistic` + `h1_event` + `h1_event_progress`, then stamps `h1_season.last_updated`.
 - **Live polling:** `useLiveData` hook (`src/shared/hooks/useLiveData.mjs`) polls `GET /api/h1/live` every 10 seconds via `setInterval` + `fetch`. A `visibilitychange` listener fires an immediate poll on tab focus. Tri-state status: `'polling'` (request in flight), `'live'` (last poll succeeded), `'offline'` (last poll failed or PWA offline). Module-level singleton ensures one connection per tab. BroadcastChannel leader election for Web Notifications.
 - **Stale version auto-reload:** Three layers detect stale client code after deployments and hard-reload: (1) Next.js `deploymentId` in `next.config.mjs` triggers hard navigation on version mismatch during client-side routing; (2) `/api/h1/live` includes `appVersion` — `useLiveData` compares it against the build-time version and reloads on mismatch (~10s detection); (3) global `ChunkLoadError` handler in `instrumentation-client.js` catches failed dynamic imports. All layers share `guardedReload()` (`src/shared/utils/reloadGuard.mjs`) — a localStorage-backed circuit breaker with 30s TTL and max 3 attempts to prevent infinite reload loops.
