+++ src/app/api/h1/rebroadcast/route.js
@@ -19,12 +19,10 @@ import { EVENT_TYPE, EVENT_STATUS } from '@/shared/enums/events.mjs';
 import { groupStatusByBucket } from '@/shared/utils/bucketing.mjs';
 
 export async function POST(request) {
-    //0. initialize
     const start = performance.now();
     let check = null;
     let formValues = null;
 
-    //0.5 validate API key
     const { error: keyError } = await validateApiKey(request);
     if (keyError === API_KEY_ERROR.DISABLED) {
         return errorResponse(403, start, 'Forbidden');
@@ -33,14 +31,12 @@ export async function POST(request) {
         return errorResponse(401, start, 'Unauthorized');
     }
 
-    //1. test if valid POST request
     const contentType = request.headers.get('content-type') || '';
     check = isValidContentType.safeParse(contentType);
     if (!check.success) {
         return errorResponse(400, start, 'Invalid Content Type');
     }
 
-    //2. get FormData and convert it to an object
     const { data: formData, error: formError } = await tryCatch(request.formData());
     if (formError) return errorResponse(400, start, 'Invalid request body');
     formValues = Object.fromEntries(formData.entries());
@@ -49,7 +45,6 @@ export async function POST(request) {
         return errorResponse(400, start, 'No action set');
     }
 
-    //3. validate FormData object structure using Zod
     check = isValidFormData.safeParse(formValues);
     if (!check.success) {
         const code = check?.error?.issues[0]?.code || null;
@@ -81,7 +76,6 @@ export async function POST(request) {
         );
     });
 
-    //4. reconstruct the HD1 wire format from the normalized tables
     let data = undefined;
     switch (formValues.action) {
         case 'get_campaign_status': {
@@ -119,11 +113,9 @@ export async function POST(request) {
             break;
     }
 
-    //5. validate data from DB
     if (data === undefined || data === null) {
         return errorResponse(404, start, 'Not found');
     }
-    //6. return response
     return successResponse(200, start, data);
 }
