+++ src/app/docs/_diagram.mjs
