+++ src/app/docs/notifications/inAppFlowDefinition.mjs
@@ -1,4 +1,4 @@
-import { buildMermaidDefinition } from '@/shared/utils/diagram.mjs';
+import { buildMermaidDefinition } from '@/app/docs/_diagram.mjs';
 
 const BODY = `
     %% Server write path
