+++ src/app/docs/notifications/pushFlowDefinition.mjs
@@ -1,4 +1,4 @@
-import { buildMermaidDefinition } from '@/shared/utils/diagram.mjs';
+import { buildMermaidDefinition } from '@/app/docs/_diagram.mjs';
 
 const BODY = `
     post_update["POST /api/h1/update"]
