+++ src/shared/utils/reloadGuard.mjs
@@ -4,22 +4,20 @@ const MAX_RELOADS = 3;
 
 export function guardedReload(reason) {
     if (typeof window === 'undefined') return;
+
+    let prevAttempts = 0;
     const raw = localStorage.getItem(GUARD_KEY);
     if (raw) {
         const [, ts, count] = raw.split(':');
         const elapsed = Date.now() - parseInt(ts, 10);
-        const attempts = parseInt(count, 10) || 0;
-        if (elapsed >= GUARD_TTL) {
-            localStorage.removeItem(GUARD_KEY);
-        } else if (attempts >= MAX_RELOADS) {
-            return;
-        } else {
-            localStorage.setItem(GUARD_KEY, `${reason}:${Date.now()}:${attempts + 1}`);
+        if (elapsed < GUARD_TTL) {
+            prevAttempts = parseInt(count, 10) || 0;
+            if (prevAttempts >= MAX_RELOADS) return;
         }
+        // else: TTL elapsed — treat as a fresh attempt window
     }
-    if (!localStorage.getItem(GUARD_KEY)) {
-        localStorage.setItem(GUARD_KEY, `${reason}:${Date.now()}:1`);
-    }
+
+    localStorage.setItem(GUARD_KEY, `${reason}:${Date.now()}:${prevAttempts + 1}`);
     window.location.reload();
 }
