+++ src/shared/utils/umami.mjs
@@ -1,3 +1,10 @@
+import { tryCatch } from '@/shared/utils/tryCatch.mjs';
+
+// Production hostname used for Umami session hashing. Inlined because
+// sendUmamiEvent early-returns in non-production, so the dev/staging
+// branches the old getHostname() switch carried were unreachable.
+const UMAMI_HOSTNAME = 'helldivers.bot';
+
 /**
  * Server-side Umami event sender. Posts directly to the Umami instance
  * (server-to-server, not subject to ad blockers). Production-only.
@@ -10,28 +17,29 @@
 async function sendUmamiEvent(payload) {
     if (process.env.NODE_ENV !== 'production') return;
 
-    await fetch(`https://${process.env.UMAMI_SITE_URL}/api/send`, {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            'User-Agent':
-                'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/137.0.0.0 Safari/537.36',
-        },
-        body: JSON.stringify({
-            type: 'event',
-            payload: {
-                website: process.env.UMAMI_SITE_ID,
-                hostname: getHostname(),
-                screen: '1x1',
-                language: 'en',
-                ...payload,
+    const { error } = await tryCatch(
+        fetch(`https://${process.env.UMAMI_SITE_URL}/api/send`, {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+                'User-Agent':
+                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/137.0.0.0 Safari/537.36',
             },
+            body: JSON.stringify({
+                type: 'event',
+                payload: {
+                    website: process.env.UMAMI_SITE_ID,
+                    hostname: UMAMI_HOSTNAME,
+                    screen: '1x1',
+                    language: 'en',
+                    ...payload,
+                },
+            }),
         }),
-    })
-        .then((response) => response.text())
-        .catch((error) => {
-            console.error('Error:', error);
-        });
+    );
+    if (error) {
+        console.error('[umami] sendUmamiEvent failed:', error?.message ?? error);
+    }
 }
 
 /**
@@ -48,18 +56,4 @@ async function sendUmamiEvent(payload) {
  */
 export async function umamiTrackEvent(title, url, name, data = {}) {
     await sendUmamiEvent({ title, url, name, data });
-}
-
-/** Maps NODE_ENV to the site hostname for Umami's session hashing. */
-function getHostname() {
-    switch (process.env.NODE_ENV) {
-        case 'development':
-            return 'localhost';
-        case 'staging':
-            return 'staging.helldivers.bot';
-        case 'production':
-            return 'helldivers.bot';
-        default:
-            throw new Error('Unknown NODE_ENV');
-    }
 }
